requirements.txt
amqp==2.5.2
awscli==1.16.314
billiard==3.6.1.0
boto3==1.10.50
botocore==1.13.50
celery==4.3.0
django-extensions==2.2.5
django-filter==2.2.0
django-model-utils==4.0.0
django-storages==1.8.0
Django==3.0.3
djangorestframework==3.11.0
gunicorn==20.0.4
ipython==7.11.1
kombu==4.6.7
mysqlclient==1.4.6
python-dotenv==0.10.3
redis==3.3.11
scrapy==1.8.0
vine==1.3.0
